+++ UserRequestApp/SubmitRequestForm.cs
@@ -167,24 +167,13 @@ namespace SinclairCC.MakeMeAdmin
                 if (Settings.RequireAuthenticationForPrivileges)
                 {
                     authenticationSuccessful = false;
-
-                    System.Net.NetworkCredential credentials = null;
-                    int authenticationReturnCode = 0;
-                    WindowsIdentity currentIdentity = WindowsIdentity.GetCurrent();
                     try
                     {
-                        do
+                        using (WindowsIdentity currentIdentity = WindowsIdentity.GetCurrent())
                         {
-                            do
-                            {
-                                credentials = NativeMethods.GetCredentials(this.Handle, currentIdentity.Name, authenticationReturnCode);
-                            } while ((null != credentials) && (string.Compare(credentials.UserName, currentIdentity.Name, true) != 0));
-
-                            if (null != credentials)
-                            {
-                                authenticationReturnCode = NativeMethods.ValidateCredentials(credentials);
-                            }
-                        } while ((null != credentials) && (authenticationReturnCode != 0));
+                            PasswordAuthenticator passwordAuthenticator = new PasswordAuthenticator();
+                            authenticationSuccessful = passwordAuthenticator.AuthenticateCurrentUser(this.Handle, currentIdentity);
+                        }
                     }
                     catch (ArgumentException excep)
                     {
@@ -198,9 +187,6 @@ namespace SinclairCC.MakeMeAdmin
                     {
                         MessageBox.Show(this, string.Format("{0}: {1}", excep.GetType().Name, excep.Message), Properties.Resources.ApplicationName, MessageBoxButtons.OK, MessageBoxIcon.Warning, MessageBoxDefaultButton.Button1, 0);
                     }
-
-                    authenticationSuccessful = (null != credentials);
-                    authenticationSuccessful &= (authenticationReturnCode == 0);
                 }
                 return authenticationSuccessful;
             }